factor out customized AppBar to is-appbar component
