+++ src/components/layout.js
@@ -4,33 +4,7 @@ import Helmet from "react-helmet"
 import "fontsource-open-sans"
 import "./layout.module.css"
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { makeStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-  },
-  appBar: {
-    color: theme.palette.secondary.main,
-    backgroundColor: theme.palette.primary.main
-  },
-   menuButton: {
-    color: "inherit"
-  },
-  title: {
-    color: "inherit"
-  },
-  testButton: {
-    color: "inherit"
-  },
-}));
+import ISAppBar from "./is-appbar";
 
 const NamedDefault = ({ children }) => {
   const data = useStaticQuery(graphql`
@@ -43,25 +17,14 @@ const NamedDefault = ({ children }) => {
       }
     }
   `)
-  const classes = useStyles();
 
   return <>
   <Helmet>
     <title>{data.site.siteMetadata.title}</title>
     <meta name="description" content={data.site.siteMetadata.description}/>
   </Helmet>
-  <CssBaseline />
-  <AppBar position="fixed" className={classes.appBar}>
-    <Toolbar>
-      <IconButton edge="start" className={classes.menuButton} aria-label="menu">
-        <MenuIcon />
-      </IconButton>
-      <Typography variant="h6" className={classes.title}>
-        News
-      </Typography>
-      <Button className={classes.testButton}>test</Button>
-    </Toolbar>
-  </AppBar>
+  <CssBaseline/>
+  <ISAppBar/>
   {children}
   </>
 }
